Extract callback handler behaviour from Membrane.Pipeline
