+++ lib/membrane/pipeline.ex
@@ -14,14 +14,14 @@ defmodule Membrane.Pipeline do
     Core,
     Element,
     Pad,
-    PlaybackState,
-    ParentSpec
+    PlaybackState
   }
 
   alias Core.Parent
   alias Core.Message
   alias Core.Pipeline.State
   alias Parent.SpecController
+  alias Membrane.Core.CallbackHandler
   import Membrane.Helper.GenServer
   require Element
   require PlaybackState
@@ -29,7 +29,6 @@ defmodule Membrane.Pipeline do
   require Pad
   use Bunch
   use Membrane.Log, tags: :core
-  use Membrane.Core.CallbackHandler
   use GenServer
 
   @typedoc """
@@ -158,7 +157,7 @@ defmodule Membrane.Pipeline do
     with {:ok, state} <-
            CallbackHandler.exec_and_handle_callback(
              :handle_init,
-             __MODULE__,
+             Core.Pipeline.ActionHandler,
              %{state: false},
              [pipeline_options],
              state
@@ -184,61 +183,6 @@ defmodule Membrane.Pipeline do
   @impl GenServer
   def terminate(reason, state) do
     :ok = state.module.handle_shutdown(reason, state.internal_state)
-  end
-
-  @impl CallbackHandler
-  # Deprecation
-  def handle_actions(%ParentSpec{} = spec, :handle_init, params, state) do
-    warn("""
-    Returning bare spec from `handle_init` is deprecated.
-    Return `{{:ok, spec: spec}, state}` instead.
-    Found in `#{inspect(state.module)}.handle_init/1`.
-    """)
-
-    super([spec: spec], :handle_init, params, state)
-  end
-
-  @impl CallbackHandler
-  def handle_actions(actions, callback, params, state) do
-    super(actions, callback, params, state)
-  end
-
-  @impl CallbackHandler
-  def handle_action(action, callback, params, state) do
-    with {:ok, state} <- do_handle_action(action, callback, params, state) do
-      {:ok, state}
-    else
-      {{:error, :invalid_action}, state} ->
-        raise CallbackError,
-          kind: :invalid_action,
-          action: action,
-          callback: {state.module, callback}
-
-      error ->
-        error
-    end
-  end
-
-  def do_handle_action({action, _args}, :handle_init, _params, state)
-      when action not in [:spec] do
-    {{:error, :invalid_action}, state}
-  end
-
-  def do_handle_action({:forward, {elementname, message}}, _cb, _params, state) do
-    Parent.Action.handle_forward(elementname, message, state)
-  end
-
-  def do_handle_action({:spec, spec = %ParentSpec{}}, _cb, _params, state) do
-    with {{:ok, _children}, state} <- SpecController.handle_spec(spec, state),
-         do: {:ok, state}
-  end
-
-  def do_handle_action({:remove_child, children}, _cb, _params, state) do
-    Parent.Action.handle_remove_child(children, state)
-  end
-
-  def do_handle_action(_action, _callback, _params, state) do
-    {{:error, :invalid_action}, state}
   end
 
   defmacro __using__(_) do
